PubsubIO: move to GCP module
